requirements.txt
# Python dependencies for OSS-Fuzz vulnerability repository scripts
PyYAML>=6.0
requests>=2.28.0
jsonschema>=4.0.0
